26.	Write a C program to check whether a number is Prime number or not.
The program takes a positive integer from the user and checks whether it is prime or not. First, it checks whether the number is 0 or 1, which is not considered prime. If the number is greater than 1, the program uses a for loop to check whether the number is divisible by any integer between 2 and half of the number. If the number is divisible by any integer in this range, it is not prime, and the program sets a flag to 1. Otherwise, if the flag remains 0, the number is prime.