27.	Write a C program to check whether a number is Armstrong number or not.
Here's how this program works:

The user is prompted to enter an integer.
The program counts the number of digits in the integer using a while loop.
The program then calculates the result by raising each digit to the power of the number of digits using a for loop and a temporary variable.
Finally, the program checks whether the result is equal to the original number. If it is, the number is an Armstrong number; otherwise, it is not.
Note: This program uses a temporary variable to avoid using the pow function.